requirements.txt
BeautifulSoup==3.2.1
Django==1.7.5
Jinja2==2.7.3
Markdown==2.6
MarkupSafe==0.23
Pillow==2.7.0
PyJWT==0.4.1
Pygments==2.0.2
boto==2.36.0
bpython==0.13.2
cffi==0.8.6
cryptography==0.7.2
cssselect==0.9.1
django-appconf==0.6
django-compressor==1.4
django-forms-bootstrap==3.0.1
django-jinja==1.1.0
django-markupfield==1.3.0
django-nose==1.3
django-smtp-ssl==1.0
django-storages==1.1.8
django-tastypie-swagger==0.1.2
django-tastypie==0.12.1
easy-thumbnails==2.2
enum34==1.0.4
gunicorn==19.1.1
lxml==3.4.1
ndg-httpsclient==0.3.3
nose==1.3.4
oauthlib==0.7.2
pbdeploy==1.8
psutil==2.1.0
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
python-cjson==1.1.0
python-dateutil==2.4.0
python-memcached==1.53
python-mimeparse==0.1.4
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2014.10
raven==5.1.1
requests-oauthlib==0.4.2
requests==2.5.1
six==1.9.0
sorl-thumbnail==12.2
wsgiref==0.1.2
